refactor: screen에서 위젯 분리
